refactor: put the evidence within reach, and let the employer line read
Two findings from walking the page as a recruiter, an engineering manager and
someone who just wanted to know what I do.

Currently sat inside the opening: the label, the portrait and three items, 513px
on a phone, between the claim and the first thing a reader could open. That put
the first openable link 48% down the page. It now sits after the short list, so
the hero is 734px instead of 1327, the work starts at 814 instead of 1407, and
the first link arrives at 33%. The opening ends on one screen. Nothing was cut
and the page is the same height; the block only moved.

It goes after the short list rather than straight after the case studies so the
two evidence sections stay adjacent, and so the reserved slot keeps its place.

The other one: the four facts a recruiter scans for were the smallest text on
the screen, set in the same grey as everything around them. The employer line
is the most legible sentence on the page for that reader, since it names where,
what the company is, and how big. It is foreground now, 17.10:1 dark and
16.02:1 light. The order is unchanged, because moving the facts above the story
turns the opening into a CV header.

LHCI clean on every blocking assertion.
